+++ backend/src/main/java/com/aicollab/backend/auth/jwt/JwtTokenProvider.java
@@ -1,0 +1,50 @@
+package com.aicollab.backend.auth.jwt;
+
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.security.Key;
+import java.util.Date;
+
+@Component
+public class JwtTokenProvider {
+
+    private final Key key;
+
+    public JwtTokenProvider(@Value("${jwt.secret}") String secretKey) {
+        this.key = Keys.hmacShaKeyFor(secretKey.getBytes());
+    }
+
+    public String createAccessToken(Long userId, String login) {
+        long now = System.currentTimeMillis();
+        long validity = 1000L * 60 * 60 * 24; // 24시간
+
+        return Jwts.builder()
+                .setSubject(String.valueOf(userId))
+                .claim("login", login)
+                .setIssuedAt(new Date(now))
+                .setExpiration(new Date(now + validity))
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public Long getUserId(String token) {
+        return Long.parseLong(
+                Jwts.parserBuilder().setSigningKey(key).build()
+                        .parseClaimsJws(token)
+                        .getBody()
+                        .getSubject()
+        );
+    }
+
+    public boolean validate(String token) {
+        try {
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+}
