--- backend/src/main/java/com/aicollab/backend/global/security/JwtTokenProvider.java
@@ -1,70 +1,0 @@
-package com.aicollab.backend.global.security;
-
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-
-@Component
-public class JwtTokenProvider {
-    private final Key key;
-    private final long accessTokenExpirationMillis;
-
-    public JwtTokenProvider(
-            @Value("${jwt.secret}") String secret,
-            @Value("${jwt.access-token-expiration-millis}") long accessTokenExpirationMillis
-    ) {
-        this.key = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
-        this.accessTokenExpirationMillis = accessTokenExpirationMillis;
-    }
-
-    public String createAccessToken(Long userId, String login) {
-        Date now = new Date();
-        Date expiry = new Date(now.getTime() + accessTokenExpirationMillis);
-
-        return Jwts.builder()
-                .setSubject(String.valueOf(userId))
-                .claim("login", login)
-                .setIssuedAt(now)
-                .setExpiration(expiry)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    public Long getUserId(String token) {
-        Claims claims = Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-
-        return Long.parseLong(claims.getSubject());
-    }
-
-    public String getLogin(String token) {
-        Claims claims = Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-
-        Object login = claims.get("login");
-        return login != null ? login.toString() : null;
-    }
-}
